[Feat] 환불 내역 조회 및 환불 가능 주문 조회 구현

+++ src/main/java/com/nhnacademy/heukbaekbookshop/order/dto/response/TossPaymentApprovalResponse.java
@@ -1,9 +1,11 @@
 package com.nhnacademy.heukbaekbookshop.order.dto.response;
 
+import java.math.BigDecimal;
+
 public record TossPaymentApprovalResponse(
         String paymentKey,
         String requestedAt,
         String approvedAt,
-        int totalAmount,
+        BigDecimal totalAmount,
         String method
 ) {}
